Added button listener for the recompute button, running into some problems now.

+++ WAVE/audit/RLA.py
@@ -55,7 +55,7 @@ class RLA(audit.Audit):
 
     def set_parameters(self, param):
         print(param)
-        self._tolerance = int(param[0])
+        self._tolerance = float(param[0])
         print(self._tolerance)
 
     def compute(self, ballot):
